+++ src/References/RegistrationActions/RegistrationActionEntry.php
@@ -1,13 +1,12 @@
 <?php
 
+declare(strict_types = 1);
+
 namespace AvtoDev\StaticReferences\References\RegistrationActions;
 
 use Illuminate\Support\Str;
 use AvtoDev\StaticReferences\References\AbstractReferenceEntry;
 
-/**
- * Сущность типа "Регистрационное действие".
- */
 class RegistrationActionEntry extends AbstractReferenceEntry
 {
     /**
@@ -27,7 +26,7 @@ class RegistrationActionEntry extends AbstractReferenceEntry
     /**
      * {@inheritdoc}
      */
-    public function configure($input = [])
+    public function configure($input = []): void
     {
         if (\is_array($input)) {
             foreach ($input as $key => $value) {
@@ -54,7 +53,7 @@ class RegistrationActionEntry extends AbstractReferenceEntry
      *
      * @return string
      */
-    public function getDescription()
+    public function getDescription(): string
     {
         return $this->description;
     }
@@ -64,7 +63,7 @@ class RegistrationActionEntry extends AbstractReferenceEntry
      *
      * @return int[]
      */
-    public function getCodes()
+    public function getCodes(): array
     {
         return $this->codes;
     }