+++ tests/bootstrap.php
@@ -1,3 +1,5 @@
 <?php
 
-new \AvtoDev\StaticReferences\Tests\Bootstrap\TestsBootstraper;
+ini_set('error_reporting', E_ALL);
+ini_set('display_errors', '1');
+ini_set('display_startup_errors', '1');